move to numpy and sycl_usm folders
