+++ onedal/datatypes/sycl_usm/data_conversion.hpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
-* Copyright 2021 Intel Corporation
+* Copyright 2024 Intel Corporation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -18,18 +18,23 @@
 
 #define PY_ARRAY_UNIQUE_SYMBOL ONEDAL_PY_ARRAY_API
 
-#include <pybind11/numpy.h>
 #include <pybind11/pybind11.h>
-#include <Python.h>
 #include <numpy/arrayobject.h>
 
 #include "oneapi/dal/table/common.hpp"
 
-namespace oneapi::dal::python {
+namespace oneapi::dal::python::sycl_usm {
 
 namespace py = pybind11;
 
-PyObject *convert_to_pyobject(const dal::table &input);
-dal::table convert_to_table(py::object inp_obj, py::object queue = py::none());
+// Convert oneDAL table with zero-copy by use of `__sycl_usm_array_interface__` protocol.
+dal::table convert_to_table(py::object obj);
 
-} // namespace oneapi::dal::python
+// Create a dictionary for `__sycl_usm_array_interface__` protocol from oneDAL table properties.
+py::dict construct_sua_iface(const dal::table& input);
+
+// Adding `__sycl_usm_array_interface__` attribute to python oneDAL table, that representing
+// USM allocations.
+void define_sycl_usm_array_property(py::class_<dal::table>& t);
+
+} // namespace oneapi::dal::python::sycl_usm
